Fix Warnings #706 (#836)
* Fix Warnings #706

Execute the following clean-up actions:

Convert control statement bodies to block
Combine nested 'if' statement in 'else' block to 'else if'
Simplify lambda expression and method reference syntax
Pull up assignment
Use instanceof keyword instead of Class.isInstance()
Exit loop earlier
Replace String concatenation by StringBuilder
Convert StringBuffer to StringBuilder for local variables
Use String.replace() instead of String.replaceAll() when possible
String.isBlank() rather than String.strip().isEmpty()
Use lazy logical operator (&& and ||)
Primitive comparison
Primitive parsing
Primitive serialization
Primitive type rather then wrapper class
Precompile reused regular expressions
Remove unnecessary String creation
Use boolean literals instead of Boolean.TRUE/FALSE when used as a primitive
Remove unused imports
Remove unused private methods
Remove unused private constructors
Remove unused private types
Remove unused private fields
Add missing '@Override' annotations
Add missing '@Override' annotations to implementations of interface methods
Add missing '@Deprecated' annotations
Remove redundant modifiers
Remove redundant semicolons
Create array with curly
Remove useless return
Remove useless continue
Convert loop into if
Remove unnecessary '$NON-NLS$' tags
Organize imports
Format source code
Remove trailing white spaces on all lines
Remove redundant String.substring() parameter
Use Arrays.fill() when possible
Evaluate without null check
Boolean value rather than comparison
Avoid double negation
Remove overridden assignment
Remove redundant comparison statement
Remove unreachable block
Operate on Maps directly
Initialize collection at creation
Initialize map at creation
Factorize operands
Single 'if' statement rather than duplicate blocks that fall through
Remove redundant if condition
Use Comparator.comparing()
Use try-with-resource
Multi-catch
Use Java method instead of system property 'FILE_ENCODING'
Use Java method instead of system property 'PATH_SEPARATOR'
Use Java method instead of system property 'FILE_SEPARATOR'
Use Java method instead of system property 'LINE_SEPARATOR'
Use Java method instead of system property 'BOOLEAN_PROPERTY'
Use diamond operator
Use Objects.hash()
Use Objects.equals() in the equals method implementation
Use Autoboxing



* Maybe these changes came from the recent upstream fetch.

* web.xml was rebuilt.  Not sure what changed the target file.

* Revert "web.xml was rebuilt.  Not sure what changed the target file."

This reverts commit 7fc2236064887aaa224047ec61c83975b1b8571b.

* Revert "Revert "web.xml was rebuilt.  Not sure what changed the target file.""

This reverts commit 4f6a208f35ad362ffbe0c0178664d1e6aab3679d.

* Revert "Fix Warnings #706"

This reverts commit 01ad9a6831a13e01c19bc02dc04a76db3c94455d.

* Revert "Revert "Revert "web.xml was rebuilt.  Not sure what changed the target file."""

This reverts commit 651999652a0e7c8e65d2d73b21ea0815366e4685.

* Fix Warnings #706

Execute the following clean-up actions (same as commit
01ad9a6831a13e01c19bc02dc04a76db3c94455d but without StringBuilder
clean-up)

+++ data/org.eclipse.birt.data.oda.pojo/src/org/eclipse/birt/data/oda/pojo/util/DataTypeUtil.java
@@ -1,12 +1,12 @@
 /*******************************************************************************
  * Copyright (c) 2013 Actuate Corporation.
- * 
+ *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * https://www.eclipse.org/legal/epl-2.0/.
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
- * 
+ *
  *
  * Contributors:
  *  Actuate Corporation  - initial API and implementation
@@ -39,7 +39,7 @@ public class DataTypeUtil {
 			return null;
 		}
 		try {
-			return Float.valueOf(NumberFormat.getInstance(locale).parse(s).floatValue());
+			return NumberFormat.getInstance(locale).parse(s).floatValue();
 		} catch (ParseException e) {
 			throw new OdaException(e);
 		}
@@ -61,7 +61,7 @@ public class DataTypeUtil {
 			return null;
 		}
 		try {
-			return Short.valueOf(NumberFormat.getInstance(locale).parse(s).shortValue());
+			return NumberFormat.getInstance(locale).parse(s).shortValue();
 		} catch (ParseException e) {
 			throw new OdaException(e);
 		}
@@ -72,7 +72,7 @@ public class DataTypeUtil {
 			return null;
 		}
 		try {
-			return Byte.valueOf(NumberFormat.getInstance(locale).parse(s).byteValue());
+			return NumberFormat.getInstance(locale).parse(s).byteValue();
 		} catch (ParseException e) {
 			throw new OdaException(e);
 		}
@@ -83,7 +83,7 @@ public class DataTypeUtil {
 			return null;
 		}
 		try {
-			return Integer.valueOf(NumberFormat.getInstance(locale).parse(s).intValue());
+			return NumberFormat.getInstance(locale).parse(s).intValue();
 		} catch (ParseException e) {
 			throw new OdaException(e);
 		}
@@ -94,7 +94,7 @@ public class DataTypeUtil {
 			return null;
 		}
 		try {
-			return Double.valueOf(NumberFormat.getInstance(locale).parse(s).doubleValue());
+			return NumberFormat.getInstance(locale).parse(s).doubleValue();
 		} catch (ParseException e) {
 			throw new OdaException(e);
 		}
